feat: moved AddLayerButton and OpacitySlider to its own components
